ref: Decoupled package path getters from the registry target
